Adapt branch to app as members
